refactor(audit): extract crypto and hash_chain submodules
Split the audit monolith further: AES-256-GCM key lifecycle and
findings_json encryption now live in phi_scan.audit.crypto, and the
HMAC-SHA256 hash chain (ChainVerifyResult, verify_audit_chain,
_compute_row_chain_hash, _attach_chain_hash, etc.) lives in
phi_scan.audit.hash_chain.

__init__.py shrinks from 1437 lines to 520 and now contains only
schema SQL, migrations, public query/insert API, _serialize_findings,
_build_scan_event_row, and identity-collection wrappers.

Package-surface compatibility is preserved via re-exports so existing
callers and monkey-patches keep working:
  - phi_scan.audit.subprocess still resolves for patch("...subprocess.run")
  - phi_scan.audit._MIGRATIONS still patchable
  - _encrypt_findings_json, _decrypt_findings_json, _hmac_sha256,
    _row_content_for_hashing, ChainVerifyResult, verify_audit_chain,
    generate_audit_key all re-exported from the package root
  - _get_current_branch, _get_current_repository_path remain patchable
  - _collect_*_identity kept in __init__ so they resolve the patchable
    git bindings (not the submodule originals)

Add regression tests in test_audit_hardening.py:
  - PHI-field sentinel (_assert_no_raw_phi_fields) rejects file_path,
    code_context, remediation_hint keys and accepts hash-only payloads
  - package-surface contract test that locks in every symbol tests
    or callers import from phi_scan.audit

Hash-chain determinism, encryption round-trip, tamper detection, and
no-raw-PHI invariants verified end-to-end via the phi-scan CLI against
a synthetic source tree and a temporary audit database.
